// src/managed/staging/github-env.d.ts
/* eslint-disable */
// Generated by Wrangler by running `wrangler types src/managed/staging/github-env.d.ts -c managed/staging/github.json --env staging --strict-vars false --env-interface StagingGithubEnv --include-runtime false` (hash: 98bd6c6cf2fb58cea76dc97b10fc6d56)
interface __BaseEnv_StagingGithubEnv {
  ENVIRONMENT: string;
  STAGING_ENABLED: string;
  STAGING_DELIVERY_ENABLED: string;
  STAGING_GITHUB_TARGET_JSON: string;
  STAGING_APPLICATION_ID: string;
  STAGING_DESTINATION_ID: string;
  STAGING_GITHUB_PRIVATE_KEY: string;
  STAGING_GITHUB_WEBHOOK_SECRET: string;
  STAGING_UNINSTALL_HMAC_KEY: string;
  STAGING_UNINSTALL_COMMITMENT_KEY: string;
  STAGING_RECONCILIATION_HMAC_KEY: string;
  STAGING_UNINSTALLS: DurableObjectNamespace<import('./github').StagingUninstall>;
  STAGING_RECONCILIATION: Fetcher /* bugdrop-managed-reconciliation-staging */;
  STAGING_AUTHORITY: Service /* entrypoint DeliveryAuthority from bugdrop-managed-authority-staging */;
  STAGING_CONTROL: Service /* entrypoint StagingControl from bugdrop-managed-authority-staging */;
}
declare namespace Cloudflare {
  interface GlobalProps {
    mainModule: typeof import('./github');
    durableNamespaces: 'StagingUninstall';
  }
  interface Env extends __BaseEnv_StagingGithubEnv {}
}
interface StagingGithubEnv extends __BaseEnv_StagingGithubEnv {}
type StringifyValues<EnvType extends Record<string, unknown>> = {
  [Binding in keyof EnvType]: EnvType[Binding] extends string ? EnvType[Binding] : string;
};
declare namespace NodeJS {
  interface ProcessEnv extends StringifyValues<
    Pick<
      Cloudflare.Env,
      | 'ENVIRONMENT'
      | 'STAGING_ENABLED'
      | 'STAGING_DELIVERY_ENABLED'
      | 'STAGING_GITHUB_TARGET_JSON'
      | 'STAGING_APPLICATION_ID'
      | 'STAGING_DESTINATION_ID'
      | 'STAGING_GITHUB_PRIVATE_KEY'
      | 'STAGING_GITHUB_WEBHOOK_SECRET'
      | 'STAGING_UNINSTALL_HMAC_KEY'
      | 'STAGING_UNINSTALL_COMMITMENT_KEY'
      | 'STAGING_RECONCILIATION_HMAC_KEY'
    >
  > {}
}

export { StagingGithubEnv };